Important fix: store variables as double again to avoid inaccuracies with runnumber representation

+++ ResultProcessor/Event.h
@@ -13,15 +13,15 @@ public:
   Event() {};
   ~Event() {};
 
-  float weight() const { return weight_; }
-  float get(const TString &var) const;
+  double weight() const { return weight_; }
+  double get(const TString &var) const;
   
 private:
-  float weight_;
-  std::map<TString,float> vars_;
+  double weight_;
+  std::map<TString,double> vars_;
 
-  void set(const TString &var, float val);
-  void setWeight(float weight) { weight_ = weight; }
+  void set(const TString &var, double val);
+  void setWeight(double weight) { weight_ = weight; }
 };
 
 typedef std::vector<Event*> Events;
